Added an (a)ll option

+++ valkyrie.sh
@@ -70,34 +70,35 @@ ctrl_c() { echo -e; echo -e; exit 0; }
 trap ctrl_c INT
 
 echo -e
-echo -e -n "${PURPLE}Install Valkyrie ${GREEN}(y/n)? ${NC}"
+echo -e -n "${PURPLE}Install Valkyrie ${GREEN}(y/n/a)? ${NC}"
 read answer
 echo -e
 if [ "$answer" != "${answer#[Yy]}" ] ;then
+    if [ "$answer" != "${answer#[Aa]}" ] ;then answer2="y"; else answer2=""; fi
 
     # PPAs
         printf "${PURPLE}Source [Valkyrie]: ${BLUE}Add Additional PPAs${NC}"
-         echo -e -n "${GREEN} (y/n)? ${NC}"; read answer; if [ "$answer" != "${answer#[Yy]}" ] ;then
+        if [ "$answer" != "${answer#[Yy]}" ] ;then printf " ${GREEN}(y/n)? ${NC} "; read answer2; else echo; fi
+        if [ "$answer2" != "${answer2#[Yy]}" ] ;then
             cmd "sudo add-apt-repository -y ppa:rock-core/qt4"
-         fi
+        fi
 
     # Dependencies
-         echo -e
-         printf "${PURPLE}Source [Valkyrie]: ${BLUE}Install Dependencies${NC}"
-         echo -e -n "${GREEN} (y/n)? ${NC}"; read answer; if [ "$answer" != "${answer#[Yy]}" ] ;then
+        echo -e
+        printf "${PURPLE}Source [Valkyrie]: ${BLUE}Install Dependencies${NC}"
+        if [ "$answer" != "${answer#[Yy]}" ] ;then printf " ${GREEN}(y/n)? ${NC} "; read answer2; else echo; fi
+        if [ "$answer2" != "${answer2#[Yy]}" ] ;then
             #libqtcore4 - might be needed to actually run if it compiles
             cmd "sudo apt install libqt4-dev"
-         fi
+        fi
 
     # Create Directories
-         echo -e
-         printf "${PURPLE}Source [Valkyrie]: ${BLUE}Create Temp Directories${NC}\n"
-         if [ -d "./src/valkyrie_tmp" ] ;then
-            printf "${BLUE}Build directory already exists, remove first? ${NC}\n"
-            printf "${YELLOW}If you leave the directoy, it will be used as-is for building.${NC}\n"
-            printf "${BLUE}Remove Directory? ${GREEN}(y/n)? ${NC}"
-            read answer
-            if [ "$answer" != "${answer#[Yy]}" ] ;then
+        echo -e
+        printf "${PURPLE}Source [Valkyrie]: ${BLUE}Create Temp Directories${NC}\n"
+        if [ -d "./src/valkyrie_tmp" ] ;then
+            printf "${BLUE}Build directory already exists, remove first${NC}\n"
+            if [ "$answer" != "${answer#[Yy]}" ] ;then printf " ${GREEN}(y/n)? ${NC} "; read answer2; else echo; fi
+            if [ "$answer2" != "${answer2#[Yy]}" ] ;then
                 cmd "sudo rm -rf ./src/valkyrie_tmp"
             fi
         fi
@@ -107,13 +108,9 @@ if [ "$answer" != "${answer#[Yy]}" ] ;then
             echo -e
             
             printf "${PURPLE}Source [Valkyrie]: ${BLUE}Use provided source snapshot${NC}"
-            if [ "$answer" != "${answer#[Yy]}" ] ;then printf " ${GREEN}(y/n)? ${NC} "; read answer2; else echo -e; fi
-            
-            #printf "${PURPLE}Source [Valkyrie]: ${BLUE}Pull current Valkyrie source (will install subservsion)${NC}\n"
-            #printf "${YELLOW}Warning! Modifications were made to source files for missing includes for compile to succeed. Using the subversion copy will likely need manual source edits to work.${NC}\n"
-            #echo -e -n "${BLUE}Use current source ${GREEN} (y/n)? ${NC}"; read answer;
+            if [ "$answer" != "${answer#[Yy]}" ] ;then printf " ${GREEN}(y/n)? ${NC} "; read answer2; else echo; fi
             cmd "mkdir -pv ./src/valkyrie_tmp/build"
-            if [ "$answer" != "${answer#[Yy]}" ] ;then
+            if [ "$answer2" != "${answer2#[Yy]}" ] ;then
                 cmd "rsync -a ./src/valkyrie-src/valkyrie/ ./src/valkyrie_tmp/build/"
             else
                 cmd "sudo apt install subversion"
@@ -137,19 +134,22 @@ if [ "$answer" != "${answer#[Yy]}" ] ;then
 
         echo -e
         printf "${PURPLE}Source [Valkyrie]: ${BLUE}Run 'qmake-qt4'${NC}"
-        echo -e -n "${GREEN} (y/n)? ${NC}"; read answer; if [ "$answer" != "${answer#[Yy]}" ] ;then
+        if [ "$answer" != "${answer#[Yy]}" ] ;then printf " ${GREEN}(y/n)? ${NC} "; read answer2; else echo; fi
+        if [ "$answer2" != "${answer2#[Yy]}" ] ;then
             cmd "qmake-qt4"
         fi
         
         echo -e
         printf "${PURPLE}Source [Valkyrie]: ${BLUE}Run 'make'${NC}"
-        echo -e -n "${GREEN} (y/n)? ${NC}"; read answer; if [ "$answer" != "${answer#[Yy]}" ] ;then
+        if [ "$answer" != "${answer#[Yy]}" ] ;then printf " ${GREEN}(y/n)? ${NC} "; read answer2; else echo; fi
+        if [ "$answer2" != "${answer2#[Yy]}" ] ;then
             cmd "make"
         fi
         
         echo -e
         printf "${PURPLE}Source [Valkyrie]: ${BLUE}Run 'make install'${NC}"
-        echo -e -n "${GREEN} (y/n)? ${NC}"; read answer; if [ "$answer" != "${answer#[Yy]}" ] ;then
+        if [ "$answer" != "${answer#[Yy]}" ] ;then printf " ${GREEN}(y/n)? ${NC} "; read answer2; else echo; fi
+        if [ "$answer2" != "${answer2#[Yy]}" ] ;then
             cmd "sudo make install"
         fi
         
@@ -161,7 +161,8 @@ if [ "$answer" != "${answer#[Yy]}" ] ;then
     # Removing build files
         echo -e
         printf "${PURPLE}Source [Valkyrie]: ${BLUE}Remove './src/valkyrie_tmp'${NC}"
-        echo -e -n "${GREEN} (y/n)? ${NC}"; read answer; if [ "$answer" != "${answer#[Yy]}" ] ;then
+        if [ "$answer" != "${answer#[Yy]}" ] ;then printf " ${GREEN}(y/n)? ${NC} "; read answer2; else echo; fi
+        if [ "$answer2" != "${answer2#[Yy]}" ] ;then
             cmd "sudo rm -rf ./src/valkyrie_tmp";
         fi
 fi
